Citizen: add operator new so we can store the memory size
The memory size is stored in a MemBlock struct immediately preceding
the memory used for the object.  The MemBlock is protected by sentinels,
and there's a trailing sentinel after the Citizen to detect overrun.

Attempts to also include the file and line number of the caller of
"new" failed since that requires the use of macros, but we can't
define macros for both 'new Foo' and 'new(ptr) Foo' (placement new).