Citizen: update census functions
There's no need for census() to allocate the result on the heap, so just
have it return a vector on the stack. This means the user doesn't have to
be responsible for freeing the result.

The census(int, memId) overload is confusing --- a differently named
method is clearer (but kept the original for backward compatibility).