Citizen: add support for declaring memory usage
Because Citizens track objects rather than memory use, tracking memory
usage has to be voluntary.

Attempted to track memory use through a custom operator new and hiding the
allocation size before the Citizen, but this doesn't work because Citizens
can be created on the stack so we can't access the allocation size.